requirements.txt
streamlit
ultralytics
opencv-python-headless
torch
torchvision
Pillow
pandas
numpy
timm
google-generativeai